requirements.txt
flask==3.0.3
flask-cors==4.0.0
pillow==10.4.0
numpy==2.1.2
opencv-python-headless==4.10.0.84
lama-cleaner
# 注意：请根据你的设备安装合适的 PyTorch 版本（CPU 或 CUDA）。
# CPU 版（示例）：
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
# CUDA 版（示例，按你驱动/CUDA 版本选择）：
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
